refactor: retire time formatter service
