+++ mpvqc/injections.py
@@ -33,7 +33,6 @@ def bindings(binder: inject.Binder) -> None:
     binder.bind_to_constructor(s.ResourceService, s.ResourceService)
     binder.bind_to_constructor(s.SettingsFileService, s.SettingsFileService)
     binder.bind_to_constructor(s.StateService, s.StateService)
-    binder.bind_to_constructor(s.TimeFormatterService, s.TimeFormatterService)
     binder.bind_to_constructor(s.VersionCheckerService, s.VersionCheckerService)
