+++ mpvqc/services/__init__.py
@@ -6,7 +6,6 @@ from .application_paths import ApplicationPathsService as ApplicationPathsServic
 from .desktop import DesktopService as DesktopService
 from .file_startup import FileStartupService as FileStartupService
 from .font_loader import FontLoaderService as FontLoaderService
-from .formatter_time import TimeFormatterService as TimeFormatterService
 from .label_width_calculator import LabelWidthCalculatorService as LabelWidthCalculatorService
 from .resource import ResourceService as ResourceService
 from .settings import MISSING as MISSING
