--- mpvqc/services/formatter_time.py
@@ -1,21 +1,0 @@
-# SPDX-FileCopyrightText: mpvQC developers
-#
-# SPDX-License-Identifier: GPL-3.0-or-later
-
-from mpvqc.shared import MILLISECONDS_PER_SECOND, SECONDS_PER_HOUR
-
-
-class TimeFormatterService:
-    @staticmethod
-    def format_time_to_string(input_seconds: float, *, long_format: bool) -> str:
-        hours, remainder = divmod(input_seconds, SECONDS_PER_HOUR)
-        minutes, seconds = divmod(remainder, 60)
-        if long_format:
-            return f"{int(hours):02d}:{int(minutes):02d}:{int(seconds):02d}"
-
-        return f"{int(minutes):02d}:{int(seconds):02d}"
-
-    @staticmethod
-    def format_milliseconds_to_string(input_milliseconds: int, *, long_format: bool) -> str:
-        seconds = input_milliseconds // MILLISECONDS_PER_SECOND
-        return TimeFormatterService.format_time_to_string(seconds, long_format=long_format)
